// src/app/modules/master-forms/Models/BranchModel.ts
export class BranchList {
    BranchId: number = 0;
    BranchCode: string = '';
    BranchName: string = '';
    BranchAddress: string = '';
    City: string = '';
    Pin: string = '';
    ContactNo: string = '';
    Email: string = '';
    Pan: string = '';
    GSTNo: string = '';
    IsActive: string = '';
    Addedby: string = '';
    CityName: string = '';
    AddedOn: string = '';
    LastUpdatedOn: string = '';
    Action: string = '';
    StateCode:string = '';
    StateName:string="";
}